updated async getCalendar method to have a faster filter loop

+++ GestionReclamation/src/main/java/tn/esprit/gestionreclamation/dto/CalendarResponse.java
@@ -13,5 +13,5 @@ import java.time.LocalDate;
 @NoArgsConstructor
 public class CalendarResponse {
     LocalDate day;
-    Integer value;
+    Long value;
 }
